Fix a bug of `.`.
When typing a `.` in shell-mode (or prolog-inferior-mode), it won't show `.` in the company-mode's candidates. This because when `company-yasnippet-autoparens` / `company-yasnippet-autoparens-2` receive 'prefix command, it call `company-grab-symbol` to grab a symbol before the point, but if that the character before the point is `.`, `company-grab-symbol` return "". Then `company--multi-backend-adapter-candidates` will filter out the "", because it does not match the prefix argument (which is "."). The workaround is let the 'prefix command carry a prefix argument.